requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
asyncpg==0.30.0
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
fastapi-mail==1.4.1
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.3
pydantic-settings==2.12.0
pydantic_core==2.41.4
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
